requirements.txt
django==4.2.7
django-widget-tweaks==1.5.0
django-jenkins==0.110.0
django[bcrypt]
pytest-django==4.5.2
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
django-recaptcha==3.0.0
django-session-timeout==0.1.0